api-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psycopg2-binary==2.9.9
tenacity==8.2.3
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
openai==1.12.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
ruff==0.1.9

